js/xpconnect/tests/mochitest/test_bug628410.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=628410
-->
<head>
  <title>Test for Bug 628410</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=628410">Mozilla Bug 628410</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>

<pre id="test">
<script type="application/javascript">

/** Test for Bug 628410 **/
{
  // window.toSource() will throw if SpecialPowers is defined on the window and
  // permissive COWs are not enabled for the global that owns its frame message
  // manager.
  let sp = window.SpecialPowers;
  window.SpecialPowers = null;

  window.toSource();
  window.SpecialPowers = sp;
}

window.toString();
InstallTrigger + "";
console + "";
ok(true, "Things didn't throw");

</script>
</pre>
</body>
</html>
